Add files via upload
This project is a Python-based QR Code Generator and Scanner application that demonstrates the practical use of QR technology. It allows users to generate a QR code from a given URL or text and save it as an image file. The generated QR code can be easily shared or displayed for scanning purposes. The project also includes a QR code scanner that uses a webcam to detect and decode QR codes in real time. OpenCV is used for camera access and QR code detection, while the qrcode library handles QR code generation. Once a QR code is successfully scanned, the embedded information is extracted. If the decoded data is a web link, it is automatically opened in the default browser. The application is simple, efficient, and user friendly. It demonstrates basic image processing and computer vision concepts. This project is suitable for beginners learning Python and real-world applications.